pyproject.toml
[project]
name = "dreamer4"
version = "0.1.0"
requires-python = "==3.10.*"
dependencies = [
    "aiohttp==3.13.3",
    "pillow==12.0.0",
    "glfw==2.10.0",
    "gpustat==1.1.1",
    "lpips==0.1.4",
    "numpy==1.24.4",
    "tensordict==0.10.0",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "torchrl==0.10.0",
    "transformers==4.56.2",
    "wandb==0.22.1",
    "pytorch-lightning==2.4.0",
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "dm-control>=1.0.0",
    "mujoco>=3.0.0",
    "jax<0.6.0",
    "jaxlib<0.6.0",
    "gymnasium[atari,accept-rom-license]",
    "envpool",
]

[tool.uv]
package = false
# glew y glib son librerías C del sistema, no se instalan con uv.
# En clusters SLURM suelen estar disponibles como módulos:
#   module load glew glib
# Si no, comprueba si ya están: ldconfig -p | grep -E "glew|glib"

